projects/s1e4/collectable-cards-api/card-image-api/src/requirements.txt
autopep8==2.0.2
boto3==1.26.110
botocore==1.29.110
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
jmespath==1.0.1
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
Pillow==9.5.0
pycodestyle==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.15
click==8.1.3
Flask==2.2.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.2.3